refactor: move ruleset into linter definition
