+++ sqlmesh/core/linter/rule.py
@@ -2,9 +2,6 @@ from __future__ import annotations
 
 import abc
 
-import operator as op
-from collections.abc import Iterator, Iterable, Set, Mapping, Callable
-from functools import reduce
 
 from sqlmesh.core.model import Model
 
@@ -48,50 +45,3 @@ class RuleViolation:
 
     def __repr__(self) -> str:
         return f"{self.rule.name}: {self.violation_msg}"
-
-
-class RuleSet(Mapping[str, type[Rule]]):
-    def __init__(self, rules: Iterable[type[Rule]] = ()) -> None:
-        self._underlying = {rule.name: rule for rule in rules}
-
-    def check_model(self, model: Model) -> t.List[RuleViolation]:
-        violations = []
-
-        for rule in self._underlying.values():
-            violation = rule().check_model(model)
-
-            if violation:
-                violations.append(violation)
-
-        return violations
-
-    def __iter__(self) -> Iterator[str]:
-        return iter(self._underlying)
-
-    def __len__(self) -> int:
-        return len(self._underlying)
-
-    def __getitem__(self, rule: str | type[Rule]) -> type[Rule]:
-        key = rule if isinstance(rule, str) else rule.name
-        return self._underlying[key]
-
-    def __op(
-        self,
-        op: Callable[[Set[type[Rule]], Set[type[Rule]]], Set[type[Rule]]],
-        other: RuleSet,
-        /,
-    ) -> RuleSet:
-        rules = set()
-        for rule in op(set(self.values()), set(other.values())):
-            rules.add(other[rule] if rule in other else self[rule])
-
-        return RuleSet(rules)
-
-    def union(self, *others: RuleSet) -> RuleSet:
-        return reduce(lambda lhs, rhs: lhs.__op(op.or_, rhs), (self, *others))
-
-    def intersection(self, *others: RuleSet) -> RuleSet:
-        return reduce(lambda lhs, rhs: lhs.__op(op.and_, rhs), (self, *others))
-
-    def difference(self, *others: RuleSet) -> RuleSet:
-        return reduce(lambda lhs, rhs: lhs.__op(op.sub, rhs), (self, *others))
